+++ R/truncate.R
@@ -32,17 +32,16 @@
 #' @examples
 #' dem <-
 #'  system.file("extdata", "demultiplexed", package = "tidyGenR")
-#' trunc_dir <- tempdir()
+#' # single end
 #' trunc_amp(
 #'     mode_trun = "pe",
 #'     in_dir = dem,
 #'     fw_pattern = "1.fastq.gz",
 #'     rv_pattern = "2.fastq.gz",
-#'     outdir = trunc_dir,
 #'     trunc_fr = c(250, 180),
 #'     max_ee = c(3, 3)
 #' )
-#' # example 2:
+#' # paired-end
 #' data("trunc_fr")
 #' trunc_amp(
 #'     mode_trun = "se",
@@ -50,7 +49,6 @@
 #'     in_dir = dem,
 #'     fw_pattern = "1.fastq.gz",
 #'     rv_pattern = "2.fastq.gz",
-#'     outdir = trunc_dir,
 #'     trunc_fr = trunc_fr,
 #'     max_ee = 3,
 #'     trunc_q = 2
@@ -59,7 +57,7 @@
 trunc_amp <- function(loci = NULL,
                       in_dir, trunc_fr,
                       fw_pattern, rv_pattern = NULL,
-                      outdir = "truncated",
+                      outdir = tempdir(),
                       filt_name = "_F_filt.fastq.gz",
                       mode_trun = "pe", multithread = FALSE,
                       max_ee = 3, trunc_q = 2) {
@@ -123,7 +121,9 @@ trunc_amp <- function(loci = NULL,
             rownames(out) <- fws$snames
             return(out)
         })
-    setNames(filtering_output, loci)
+    z <- setNames(filtering_output, loci)
+    message("Filtered/truncated reads written to", outdir)
+    return(z)
 }
 
 #' helper functions to trunc_amp: get truncation lengths