+++ R/variant_calling.R
@@ -61,7 +61,7 @@
 #' variant_call(in_dir = truncated)
 #' @export
 variant_call <- function(loci = NULL,
-                         in_dir = NULL,
+                         in_dir,
                          fw_pattern = "_F_filt.fastq.gz",
                          rv_pattern = NULL,
                          sample_locus = "(^[a-zA-Z0-9]*).([a-zA-Z0-9]*)",