+++ tests/testthat/test-truncate.R
@@ -13,8 +13,7 @@ test_that("Default parameters work", {
             fw_pattern = "1.fastq.gz",
             rv_pattern = "2.fastq.gz",
             trunc_fr = 200,
-            max_ee = c(3, 3),
-            outdir = file.path(tempdir(), "truncated")
+            max_ee = c(3, 3)
         )
     expect_equal(names(tr), loci)
     expect_equal(class(tr), "list")
@@ -30,8 +29,7 @@ test_that("paired-end accepts arguments correctly. paired-end, all loci:", {
             fw_pattern = "1.fastq.gz",
             rv_pattern = "2.fastq.gz",
             trunc_fr = 200,
-            max_ee = c(3, 3),
-            outdir = file.path(tempdir(), "truncated")
+            max_ee = c(3, 3)
         )
     # pe retturns error when rv_pattern is not stated
     expect_error(
@@ -42,8 +40,7 @@ test_that("paired-end accepts arguments correctly. paired-end, all loci:", {
             fw_pattern = "1.fastq.gz",
             # rv_pattern = "2.fastq.gz",
             trunc_fr = 200,
-            max_ee = c(3, 3),
-            outdir = file.path(tempdir(), "truncated")
+            max_ee = c(3, 3)
         )
     )
     # max_ee of numeric(1) is equivalent to numeric(2)
@@ -55,8 +52,7 @@ test_that("paired-end accepts arguments correctly. paired-end, all loci:", {
             fw_pattern = "1.fastq.gz",
             rv_pattern = "2.fastq.gz",
             trunc_fr = 200,
-            max_ee = 3,
-            outdir = file.path(tempdir(), "truncated")
+            max_ee = 3
         )
     tr_pe_maxee2 <-
         trunc_amp(
@@ -66,8 +62,7 @@ test_that("paired-end accepts arguments correctly. paired-end, all loci:", {
             fw_pattern = "1.fastq.gz",
             rv_pattern = "2.fastq.gz",
             trunc_fr = 200,
-            max_ee = c(3, 3),
-            outdir = file.path(tempdir(), "truncated")
+            max_ee = c(3, 3)
         )
     expect_identical(tr_pe_maxee1, tr_pe_maxee2)
 
@@ -80,8 +75,7 @@ test_that("paired-end accepts arguments correctly. paired-end, all loci:", {
             fw_pattern = "1.fastq.gz",
             rv_pattern = "2.fastq.gz",
             trunc_fr = 200,
-            max_ee = 3,
-            outdir = file.path(tempdir(), "truncated")
+            max_ee = 3
         )
     tr_pe2 <-
         trunc_amp(
@@ -91,8 +85,7 @@ test_that("paired-end accepts arguments correctly. paired-end, all loci:", {
             fw_pattern = "1.fastq.gz",
             rv_pattern = "2.fastq.gz",
             trunc_fr = c(200, 200),
-            max_ee = 3,
-            outdir = file.path(tempdir(), "truncated")
+            max_ee = 3
         )
     expect_identical(tr_pe2, tr_pe1)
 })
@@ -123,8 +116,7 @@ test_that("only first numeric is taken when fed with numeric(2) trunc_fr in sing
             mode_trun = "se",
             in_dir = dem,
             fw_pattern = "1.fastq.gz",
-            trunc_fr = c(100, 200),
-            outdir = file.path(tempdir(), "truncated")
+            trunc_fr = c(100, 200)
         ))
   }
   )
@@ -138,8 +130,7 @@ test_that("not correct length of max_ee", {
             in_dir = dem,
             fw_pattern = "1.fastq.gz",
             trunc_fr = 200,
-            max_ee = c(3, 3),
-            outdir = file.path(tempdir(), "truncated")
+            max_ee = c(3, 3)
         )
     )
 })
@@ -153,9 +144,7 @@ test_that("trunc_fr dataframe feeds with locus-specific truncation lengths. trun
     in_dir = dem,
     fw_pattern = "1.fastq.gz",
     rv_pattern = "2.fastq.gz",
-    trunc_fr = trunc_fr,
-    outdir = file.path(tempdir(), "truncated")
-  )
+    trunc_fr = trunc_fr)
 
   expect_length(h, 2)  # Check if the list has exactly 2 elements
 })